backend/requirements.txt
# FastAPI
fastapi==0.115.6
uvicorn[standard]==0.34.0

# Database
sqlalchemy==2.0.36
mysql-connector-python==9.1.0

# Authentication & Security
python-jose[cryptography]==3.3.0
python-multipart==0.0.20
passlib[bcrypt]==1.7.4
cryptography==43.0.3

# Environment
python-dotenv==1.0.1

# CORS
starlette==0.41.3
